Proyecto Ejecutable y terminado

+++ src/vos/Reserva.java
@@ -5,37 +5,37 @@ import java.util.Date;
 import org.codehaus.jackson.annotate.*;
 
 /**
- * @author Nicolás Mateo Hernández Rojas 	- 	nm.hernandez10@uniandes.edu.co
- * @author David Felipe Niño Romero		-	df.nino10@uniandes.edu.co
- * Clase que representa a las Reservas del modelo AlohAndes
+ * @author Nicolás Mateo Hernández Rojas - nm.hernandez10@uniandes.edu.co
+ * @author David Felipe Niño Romero - df.nino10@uniandes.edu.co Clase que
+ *         representa a las Reservas del modelo AlohAndes
  */
 
-
-public class Reserva 
-{
+public class Reserva {
 	@JsonProperty(value = "idCliente")
 	private long idCliente;
-	
+
 	@JsonProperty(value = "idEspacio")
 	private long idEspacio;
-	
+
 	@JsonProperty(value = "fechaInicio")
 	private Date fechaInicio;
-	
+
 	@JsonProperty(value = "duracion")
 	private int duracion;
-	
+
 	@JsonProperty(value = "fechaReserva")
 	private Date fechaReserva;
-	
+
 	@JsonProperty(value = "cancelado")
 	private boolean cancelado;
-	
+
 	@JsonProperty(value = "precio")
 	private double precio;
-	
-	public Reserva(@JsonProperty(value = "idCliente") long idCliente, @JsonProperty(value = "idEspacio") long idEspacio, @JsonProperty(value = "fechaInicio") Date fechaInicio, @JsonProperty(value = "duracion") int duracion, @JsonProperty(value = "fechaReserva") Date fechaReserva, @JsonProperty(value = "cancelado")boolean cancelado, @JsonProperty(value = "precio")double precio)
-	{
+
+	public Reserva(@JsonProperty(value = "idCliente") long idCliente, @JsonProperty(value = "idEspacio") long idEspacio,
+			@JsonProperty(value = "fechaInicio") Date fechaInicio, @JsonProperty(value = "duracion") int duracion,
+			@JsonProperty(value = "fechaReserva") Date fechaReserva,
+			@JsonProperty(value = "cancelado") boolean cancelado, @JsonProperty(value = "precio") double precio) {
 		this.idCliente = idCliente;
 		this.idEspacio = idEspacio;
 		this.fechaInicio = fechaInicio;
@@ -44,113 +44,92 @@ public class Reserva
 		this.precio = precio;
 		this.cancelado = cancelado;
 	}
-	
-	public double getPrecio() 
-	{		
+
+	public double getPrecio() {
 		return precio;
 	}
 
-	public void setPrecio(double precio) 
-	{
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
 
-	public long getIdCliente() 
-	{
+	public long getIdCliente() {
 		return idCliente;
 	}
-	
-	public boolean isCancelado()
-	{
+
+	public boolean isCancelado() {
 		return cancelado;
 	}
 
-	public void setCancelado(boolean cancelado) 
-	{
+	public void setCancelado(boolean cancelado) {
 		this.cancelado = cancelado;
 	}
 
-	public void setIdCliente(long idCliente)
-	{
+	public void setIdCliente(long idCliente) {
 		this.idCliente = idCliente;
 	}
 
-	public Date getFechaReserva() 
-	{
+	public Date getFechaReserva() {
 		return fechaReserva;
 	}
 
-	public void setFechaReserva(Date fechaReserva) 
-	{
+	public void setFechaReserva(Date fechaReserva) {
 		this.fechaReserva = fechaReserva;
 	}
 
-	public long getIdEspacio()
-	{
+	public long getIdEspacio() {
 		return idEspacio;
 	}
 
-
-	public void setIdEspacio(long idEspacio)
-	{
+	public void setIdEspacio(long idEspacio) {
 		this.idEspacio = idEspacio;
 	}
 
-	public Date getFechaInicio() 
-	{
+	public Date getFechaInicio() {
 		return fechaInicio;
 	}
 
-	public void setFechaInicio(Date fechaInicio) 
-	{
+	public void setFechaInicio(Date fechaInicio) {
 		this.fechaInicio = fechaInicio;
 	}
 
-	public int getDuracion()
-	{
+	public int getDuracion() {
 		return duracion;
 	}
 
-	public void setDuracion(int duracion) 
-	{
+	public void setDuracion(int duracion) {
 		this.duracion = duracion;
-	}	
-	
-	public Date calcularFechaFin()
-	{
+	}
+
+	public Date calcularFechaFin() {
 		Date fechaFin = new Date();
 		int meses = 0;
-		if(duracion/30 > 0)
-		{
-			meses = (Integer)(duracion/30);
+		if (duracion / 30 > 0) {
+			meses = duracion / 30;
 		}
-		int dias = duracion - meses*30;
-		if(fechaInicio.getMonth() + meses > 12)
-		{
-			fechaFin.setYear(fechaInicio.getYear()+1);
-			fechaFin.setMonth(fechaInicio.getMonth()+meses-12);
+		int dias = duracion - meses * 30;
+		if (fechaInicio.getMonth() + meses > 12) {
+			fechaFin.setYear(fechaInicio.getYear() + 1);
+			fechaFin.setMonth(fechaInicio.getMonth() + meses - 12);
 			fechaFin.setDate(dias);
 		}
-		
-		return fechaFin;		
+
+		return fechaFin;
 	}
-	
-	public Date calcularFechaConDiasDespues(int diasMas)
-	{
+
+	public Date calcularFechaConDiasDespues(int diasMas) {
 		Date fechaCon = new Date();
 		int meses = 0;
-		if(diasMas/30 > 0)
-		{
-			meses = (Integer)(diasMas/30);
+		if (diasMas / 30 > 0) {
+			meses = diasMas / 30;
 		}
-		int dias = diasMas - meses*30;
-		if(fechaInicio.getMonth() + meses > 12)
-		{
-			fechaCon.setYear(fechaInicio.getYear()+1);
-			fechaCon.setMonth(fechaInicio.getMonth()+meses-12);
+		int dias = diasMas - meses * 30;
+		if (fechaInicio.getMonth() + meses > 12) {
+			fechaCon.setYear(fechaInicio.getYear() + 1);
+			fechaCon.setMonth(fechaInicio.getMonth() + meses - 12);
 			fechaCon.setDate(dias);
 		}
-		
+
 		return fechaCon;
 	}
 }
